Add a simple test with acceleration limits
