--- diff_drive_controller/test/config/test_diff_drive_controller_limits.yaml
@@ -1,42 +1,0 @@
-test_diff_drive_controller:
-  ros__parameters:
-    left_wheel_names: ["left_wheels"]
-    right_wheel_names: ["right_wheels"]
-
-    wheel_separation: 0.40
-    wheel_radius: 0.02
-
-    wheel_separation_multiplier: 1.0
-    left_wheel_radius_multiplier: 1.0
-    right_wheel_radius_multiplier: 1.0
-
-    odom_frame_id: odom
-    base_frame_id: base_link
-    pose_covariance_diagonal: [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-    twist_covariance_diagonal: [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-
-    position_feedback: false
-    open_loop: true
-    enable_odom_tf: true
-
-    cmd_vel_timeout: 0.5 # seconds
-    publish_limited_velocity: true
-    velocity_rolling_window_size: 10
-
-    linear.x.max_velocity: 10.0
-    linear.x.min_velocity: -10.0
-    linear.x.max_acceleration: 10.0
-    linear.x.max_deceleration: -20.0
-    linear.x.max_acceleration_reverse: -10.0
-    linear.x.max_deceleration_reverse: 20.0
-    linear.x.max_jerk: 50.0
-    linear.x.min_jerk: -50.0
-
-    angular.z.max_velocity: .NAN
-    angular.z.min_velocity: .NAN
-    angular.z.max_acceleration: .NAN
-    angular.z.max_deceleration: .NAN
-    angular.z.max_acceleration_reverse: .NAN
-    angular.z.max_deceleration_reverse: .NAN
-    angular.z.max_jerk: .NAN
-    angular.z.min_jerk: .NAN
